refactor: split builtins.rs into module directory for parallel development
Convert builtins.rs from a single file to a module directory with
individual files per built-in function. This enables parallel development
of new built-in functions without merge conflicts.

Part of #1049
